+++ crates/holochain/src/core/workflow/sys_validation_workflow/tests.rs
@@ -5,6 +5,7 @@ use crate::test_utils::host_fn_caller::*;
 use crate::test_utils::{assert_limbo_empty, wait_for_integration};
 use crate::{conductor::ConductorHandle, core::MAX_TAG_SIZE};
 use holo_hash::fixt::AgentPubKeyFixturator;
+use holochain_keystore::SignedActionHashedExt;
 use holochain_wasm_test_utils::TestWasm;
 use holochain_zome_types::action::SignedAction;
 use holochain_zome_types::fixt::{ActionFixturator, CreateAction, DnaAction};
