+++ crates/holochain_types/src/test_utils.rs
@@ -2,11 +2,7 @@
 
 use crate::dna::wasm::DnaWasm;
 use crate::prelude::*;
-use crate::record::SignedActionHashedExt;
-use holochain_keystore::MetaLairClient;
-use holochain_zome_types::prelude::{
-    ActionHeader, AgentValidationPkgData, CreateData, DnaData, InitZomesCompleteData, UpdateData,
-};
+use holochain_zome_types::prelude::{CreateData, UpdateData};
 pub use holochain_zome_types::test_utils::*;
 use std::path::PathBuf;
 
@@ -98,35 +94,6 @@ pub fn fake_cap_secret() -> CapSecret {
     [0; CAP_SECRET_BYTES].into()
 }
 
-/// Create a fake SignedActionHashed and EntryHashed pair with random content
-pub async fn fake_unique_record(
-    keystore: &MetaLairClient,
-    agent_key: AgentPubKey,
-    visibility: EntryVisibility,
-) -> anyhow::Result<(SignedActionHashed, EntryHashed)> {
-    let content: SerializedBytes =
-        UnsafeBytes::from(nanoid::nanoid!().as_bytes().to_owned()).into();
-    let entry = Entry::App(content.try_into().unwrap()).into_hashed();
-    let app_entry_def = AppEntryDefFixturator::new(visibility).next().unwrap();
-    let action_1 = Action {
-        header: ActionHeader {
-            author: agent_key,
-            timestamp: Timestamp::now(),
-            action_seq: 0,
-            prev_action: Some(fake_action_hash(1)),
-        },
-        data: ActionData::Create(CreateData {
-            entry_type: EntryType::App(app_entry_def),
-            entry_hash: entry.as_hash().to_owned(),
-        }),
-    };
-
-    Ok((
-        SignedActionHashed::sign(keystore, action_1.into_hashed()).await?,
-        entry,
-    ))
-}
-
 #[allow(missing_docs)]
 pub trait ActionRefMut {
     fn author_mut(&mut self) -> &mut AgentPubKey;
@@ -177,107 +144,4 @@ impl ActionRefMut for Action {
     fn timestamp_mut(&mut self) -> &mut Timestamp {
         &mut self.header.timestamp
     }
-}
-
-/// Create test chain data
-pub async fn valid_arbitrary_chain(
-    keystore: &MetaLairClient,
-    author: AgentPubKey,
-    length: usize,
-) -> Vec<Record> {
-    use ::fixt::*;
-    use holo_hash::fixt::DnaHashFixturator;
-
-    let mut out = Vec::new();
-    let extend_out = |mut out: Vec<Record>, action: Action, entry: Option<Entry>| async move {
-        out.push(Record::new(
-            SignedActionHashed::sign(keystore, action.into_hashed())
-                .await
-                .unwrap(),
-            match entry {
-                Some(entry) => RecordEntry::Present(entry),
-                None => RecordEntry::NA,
-            },
-        ));
-        out
-    };
-
-    let dna = Action {
-        header: ActionHeader {
-            author: author.clone(),
-            timestamp: Timestamp::now(),
-            action_seq: 0,
-            prev_action: None,
-        },
-        data: ActionData::Dna(DnaData {
-            dna_hash: fixt!(DnaHash),
-        }),
-    };
-    out = extend_out(out, dna, None).await;
-
-    let avp = Action {
-        header: ActionHeader {
-            author: author.clone(),
-            timestamp: Timestamp::now(),
-            action_seq: 1,
-            prev_action: Some(out.last().as_ref().unwrap().action_address().clone()),
-        },
-        data: ActionData::AgentValidationPkg(AgentValidationPkgData {
-            membrane_proof: None,
-        }),
-    };
-    out = extend_out(out, avp, None).await;
-
-    let agent_entry = Entry::Agent(author.clone());
-
-    let agent = Action {
-        header: ActionHeader {
-            author: author.clone(),
-            timestamp: Timestamp::now(),
-            action_seq: 2,
-            prev_action: Some(out.last().as_ref().unwrap().action_address().clone()),
-        },
-        data: ActionData::Create(CreateData {
-            entry_type: EntryType::AgentPubKey,
-            entry_hash: agent_entry.clone().into_hashed().hash,
-        }),
-    };
-    out = extend_out(out, agent, Some(agent_entry)).await;
-
-    let init_zomes = Action {
-        header: ActionHeader {
-            author: author.clone(),
-            timestamp: Timestamp::now(),
-            action_seq: 3,
-            prev_action: Some(out.last().as_ref().unwrap().action_address().clone()),
-        },
-        data: ActionData::InitZomesComplete(InitZomesCompleteData {}),
-    };
-    out = extend_out(out, init_zomes, None).await;
-
-    for action_seq in 4..length {
-        let entry = Entry::App(AppEntryBytes(SerializedBytes::from(UnsafeBytes::from(
-            nanoid::nanoid!().as_bytes().to_owned(),
-        ))));
-
-        let action = Action {
-            header: ActionHeader {
-                author: author.clone(),
-                timestamp: Timestamp::now(),
-                action_seq: action_seq as u32,
-                prev_action: Some(out.last().as_ref().unwrap().action_address().clone()),
-            },
-            data: ActionData::Create(CreateData {
-                entry_type: EntryType::App(AppEntryDef::new(
-                    0.into(),
-                    1.into(),
-                    EntryVisibility::Public,
-                )),
-                entry_hash: entry.clone().into_hashed().hash,
-            }),
-        };
-        out = extend_out(out, action, Some(entry)).await;
-    }
-
-    out
 }
